tama jumping

+++ platformio/tamagotchi/lib/StatsScreen/StatsScreen.cpp
@@ -28,7 +28,7 @@ StatsScreen::StatsScreen(): BaseScreen(1)
         "Weight"
     };
     for (unsigned int index = 0; index < (sizeof(bar_list) / sizeof(bar_list[0])); index++) {
-        render_list.push_back(new DisplayText{15, index * 14, bar_list[index]});
+        render_list.push_back(new DisplayText{15, short(index * 14), bar_list[index]});
     }
 }
